+++ bin/stop.sh
@@ -1,8 +1,8 @@
-#!/bin/bash
+#!/usr/bin/env bash
 #-----------------------------------------------------------#
 # @author: dep
 # @link: https://github.com/demmonico
-# @package: https://github.com/demmonico/docker-ci
+# @package: https://github.com/demmonico/docker-manager
 #
 # This script stops docker container(s) and unused networks
 # If PROJECT_NAME doesn't exists then script will stop all exists containers and unused networks
@@ -11,7 +11,10 @@
 # Params:
 #   -c - remove containers after they stops
 #   -a - remove all containers and their images after they stops
+#   -f - forced mode
 #-----------------------------------------------------------#
+
+
 
 # get params
 while [[ $# -gt 0 ]]
@@ -20,6 +23,7 @@ do
     case $key in
         -c) isRemoveContainers='true';;
         -a) isRemoveAll='true';;
+        -f) isForceMode='-f';;
         -n)
             if [ ! -z "$2" ]; then
                 export PROJECT="$2"
@@ -34,13 +38,12 @@ do
         shift
 done
 
-
 # bin dir
-DC_BIN_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+DM_BIN_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 # docker containers root dir
-DC_ROOT_DIR="$(dirname "$DC_BIN_DIR")"
+DM_ROOT_DIR="$(dirname "${DM_BIN_DIR}")"
 # dir of projects
-DC_PROJECT_DIR="$DC_ROOT_DIR/projects"
+DM_PROJECT_DIR="${DM_ROOT_DIR}/projects"
 
 
 
@@ -50,18 +53,18 @@ export GITHUB_TOKEN=""
 
 
 # one/all processing
-if [ ! -z "$PROJECT" ]
+if [ ! -z "${PROJECT}" ]
 then
     # if project exists
-    if [ -d "$DC_PROJECT_DIR/$PROJECT" ]
+    if [ -d "${DM_PROJECT_DIR}/${PROJECT}" ]
     then
-        cd "$DC_PROJECT_DIR/$PROJECT"
+        cd "${DM_PROJECT_DIR}/${PROJECT}"
         # remove all
-        if [ ! -z "$isRemoveAll" ]
+        if [ ! -z "${isRemoveAll}" ]
         then
             docker-compose down --rmi all
         # remove containers
-        elif [ ! -z "$isRemoveContainers" ]
+        elif [ ! -z "${isRemoveContainers}" ]
         then
             docker-compose down
         # only stop containers
@@ -69,28 +72,28 @@ then
             docker-compose stop
         fi
     else
-        echo "Invalid project - $PROJECT"
+        echo "Invalid project - ${PROJECT}"
     fi
 else
-    cd "$DC_ROOT_DIR"
+    cd "${DM_ROOT_DIR}"
     CONTAINERS=$(docker ps -a -q)
 
     # stop all containers
     echo "Stopped containers (id):"
-    docker stop $CONTAINERS
+    docker stop ${CONTAINERS}
 
     # remove all containers
-    if [ ! -z "$isRemoveAll" ] || [ ! -z "$isRemoveContainers" ]
+    if [ ! -z "${isRemoveAll}" ] || [ ! -z "${isRemoveContainers}" ]
     then
         echo "Removed containers (id):"
-        docker rm -f $CONTAINERS
+        docker rm ${isForceMode} ${CONTAINERS}
     fi
 
     # remove all images
-    if [ ! -z "$isRemoveAll" ]
+    if [ ! -z "${isRemoveAll}" ]
     then
         echo "Removed images (id):"
-        docker rmi $(docker images -q)
+        docker rmi ${isForceMode} $(docker images -q)
     fi
 fi;